Criacao da classe APIrequest, implementacao das funcoes adicionar e remover ativos usando APIrequest
