+++ src/main/java/org/controladorinvestimentos/controlador_investimentos/beans/controladorAtivos.java
@@ -8,51 +8,14 @@ import okhttp3.Response;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import org.controladorinvestimentos.controlador_investimentos.beans.Ativo;
-
 import java.io.IOException;
 import java.util.ArrayList;
 
 public class controladorAtivos {
 
-    public double buscarPrecoAtivoEmTempoReal(String simbolo) throws IOException {
-        String apiKey = "7kfUNQUQm5GxWV6GXAf3ig";
-        String url = "https://brapi.dev/api/quote/" + simbolo + "?token=" + apiKey;
-
-        OkHttpClient client = new OkHttpClient();
-        Request request = new Request.Builder()
-                .url(url)
-                .build();
-
-        try (Response response = client.newCall(request).execute()) {
-            if (!response.isSuccessful()) {
-                throw new IOException("Erro na requisição: " + response.code());
-            }
-
-
-            String jsonResponse = response.body().string();
-
-
-            JsonObject jsonObject = JsonParser.parseString(jsonResponse).getAsJsonObject();
-            JsonObject results = jsonObject.getAsJsonArray("results").get(0).getAsJsonObject();
-            double preco = results.get("regularMarketPrice").getAsDouble();
-
-            return preco;
-        }
-    }
-
-    public void BuscarPreco(String simbolo, Ativo ativo) {
-        try {
-            double preco = buscarPrecoAtivoEmTempoReal(simbolo);
-            ativo.setPreco(preco);
-            System.out.println("Preço atualizado para o ativo " + ativo.getNome() + ": " + preco);
-        } catch (IOException e) {
-            System.err.println("Erro ao buscar o preço do ativo: " + e.getMessage());
-        }
-    }
-
     private static iRepositorioAtivos repositorioAtivos;
 
-    private void CriarAtivo(int idAtv, double ValorAtv, String nome){
+    public static void CriarAtivo(String nome) throws IOException {
         repositorioAtivos = RepositorioAtivos.getInstance();
         try {
             Ativo _AtivoEncontrado = repositorioAtivos.buscarAtivo(nome);
@@ -61,11 +24,12 @@ public class controladorAtivos {
 
             }
         } catch (Exception e) {
-            repositorioAtivos.adicionarAtivo(nome,ValorAtv);
+            double ValorAtualAtivo = APIrequest.buscarPrecoAtivoEmTempoReal(nome);
+            repositorioAtivos.adicionarAtivo(nome,ValorAtualAtivo);
         }
     }
 
-    private void RemoverAtivo(int idAtv, double ValorAtv, String nome){
+    public static void RemoverAtivo(String nome){
         try {
             Ativo _AtivoEncontrado = repositorioAtivos.buscarAtivo(nome);
             if (_AtivoEncontrado != null) {
@@ -76,16 +40,6 @@ public class controladorAtivos {
         }
 
     }
-    private void AlterarPreco(String nome,double preco){
-        try {
-            Ativo _AtivoEncontrado = repositorioAtivos.buscarAtivo(nome);
-
-            if (_AtivoEncontrado != null) {
-                repositorioAtivos.AlterarPreco(preco,_AtivoEncontrado);
-            }
-        }catch (Exception e) {
-            throw new Exist("Ativo não existe no sistema.");
-        }
 
     }
-}
+
